feat: integrate create job
