+++ lib/actions/job/createJob.ts
@@ -19,7 +19,7 @@ export const PayloadAddJob = z.object({
   analyzeCv: z.boolean(),
 });
 
-const connectCvJob = async ({
+export const connectCvJob = async ({
   cvId,
   orgId,
   batchJobId,
@@ -41,43 +41,6 @@ const connectCvJob = async ({
     if (analyzeCvEnabled) {
       analyzeCv({ cvAnalysisId: cvAnalyze.id, jobId: batchJobId });
     }
-  } catch (error) {
-    return errorHandler(error);
-  }
-};
-
-export const createJob = async (
-  payload: z.infer<typeof PayloadAddJob>,
-  cv: FormData,
-) => {
-  try {
-    const cvData = cv.getAll('UPLOAD') as File[];
-    const cvLinkedin = cv.getAll('LINKEDIN') as string[];
-
-    const { analyzeCv, ...safePayload } = PayloadAddJob.parse(payload);
-    const job = await prismadb.batchJob.create({
-      data: {
-        ...safePayload,
-      },
-    });
-
-    console.log('Job Create Prisma', job);
-
-    const onSuccess = (props: { id: string; url: string; key: string }) => {
-      connectCvJob({
-        cvId: props.id,
-        orgId: safePayload.orgId,
-        batchJobId: job.id,
-        analyzeCvEnabled: analyzeCv,
-      });
-    };
-    cvData.forEach((cv) => {
-      uploadCv({ cv, source: 'UPLOAD', orgId: payload.orgId }, onSuccess);
-    });
-    cvLinkedin.forEach((cv) => {
-      uploadCv({ cv, source: 'LINKEDIN', orgId: payload.orgId }, onSuccess);
-    });
-    return job;
   } catch (error) {
     return errorHandler(error);
   }
